feature: image upload. Upload product image
I'm not happy with how it works though.

With the upload display in the form, when you cancel an errored upload,
it stays in the form because right after the cancel event it sends an
`allow_upload` event, which I can't find any reference to, which seems
to re-instate the upload. So you can't make a bad upload go away.

With the upload display NOT in the form, the upload preview and progress
disappear when it is done auto-uploading, with no feedback on the file
input. So you see a blank upload well, blank file input. But when you
click save, it does save your uploaded image.